back/django_server/requirements.txt
cloudinary==1.11.0
gunicorn==19.7.1
django==2.0.3
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
setuptools==39.0.1
numpy==1.14.2
psycopg2-binary==2.7.4
dj-database-url==0.5.0
